requirements.txt
face_recognition
streamlit
opencv-python
cmake
PyYAML